Refactor Layout component by extracting Header and Footer into separate components for improved organization and maintainability
